update the flow of state
